refactor: Move PoisonNeedle#doHit to AbstractDart
