+++ src/main/java/com/aetherteam/aether/entity/projectile/PoisonNeedle.java
@@ -3,15 +3,10 @@ package com.aetherteam.aether.entity.projectile;
 import com.aetherteam.aether.effect.AetherEffects;
 import com.aetherteam.aether.entity.AetherEntityTypes;
 import com.aetherteam.aether.entity.projectile.dart.AbstractDart;
-import com.aetherteam.aether.mixin.mixins.common.accessor.PlayerAccessor;
 import net.minecraft.world.effect.MobEffectInstance;
-import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.phys.EntityHitResult;
-import net.minecraft.world.phys.HitResult;
 
 public class PoisonNeedle extends AbstractDart {
     public PoisonNeedle(EntityType<? extends PoisonNeedle> type, Level level) {
@@ -22,22 +17,6 @@ public class PoisonNeedle extends AbstractDart {
     public PoisonNeedle(Level level, LivingEntity shooter) {
         super(AetherEntityTypes.POISON_NEEDLE.get(), level, shooter, null);
         this.setBaseDamage(0.25);
-    }
-
-    /**
-     * Handles shield damaging when this projectile hits an entity.
-     * @param result The {@link HitResult} of the projectile.
-     */
-    @Override
-    protected void onHit(HitResult result) {
-        super.onHit(result);
-        if (result.getType() == HitResult.Type.ENTITY) {
-            Entity entity = ((EntityHitResult) result).getEntity();
-            if (entity instanceof Player player && player.isBlocking()) {
-                PlayerAccessor playerAccessor = (PlayerAccessor) player;
-                playerAccessor.callHurtCurrentlyUsedShield(3.0F);
-            }
-        }
     }
 
     /**
